+++ API/Resources/UserResourceResult.cs
@@ -16,6 +16,7 @@
 */
 
 using Models;
+using System.Collections.Generic;
 
 namespace API.Resources
 {
@@ -63,6 +64,11 @@ namespace API.Resources
         /// Gets or sets the institution where the user is registered.
         /// </summary>
         public Institution Institution { get; set; }
+       
+        /// <summary>
+        /// Gets or sets the followed projects for user
+        /// </summary>
+        public List<UserProjectResourceResult> UserProject { get; set; }
     }
 
 }